Add ES6 Modules
