+++ src/style.css
@@ -30,6 +30,8 @@ h1 {
 .score-list {
   border: 2px solid;
   width: 25rem;
+  height: 20rem;
+  overflow:scroll;
 }
 
 input,
